+++ app-shell/src/androidTest/java/com/yokuli/marine/shell/ShellActivityStoryTest.kt
@@ -149,6 +149,7 @@ class ShellActivityStoryTest {
                                 defaultDesktopDocument,
                                 mapConfigured = false,
                                 theme = WpThemeSpec(),
+                                visualContributions = productionVisualContributions,
                             ),
                             onAction = {},
                         )